Early draft of service documentation intent.
This commit establishes the split into project workbook
and service model, and stubs out all of the relevant sections
and subsections for the model.  It adds explanatory notes
for about a third of them, with more to come later.  It also
adds explanatory annotations to the project workbook documents.

Note that Mylyn Intent is still broken for basic documents --
it's really intended to document source code artifacts, so asking
it to track and document something like an MS Project plan or
a Word document is a little twitchy right now.  But I'm hopeful,
so here it is.

Just before this commit, I used Intent to generate an HTML export
of the documentation to date -- that's in the "generated" folder.
It's not really canon content, but it's nice to have it showing
up in HTML format in the repository for ease of reference for
people who don't have Eclipse and Intent installed.